Updated README, moved NTImageView into /NTImageView
